+++ src/supervisor/watchers/handlers/replication-controller.ts
@@ -14,6 +14,7 @@ import {
   kubernetesObjectToWorkloadAlreadyScanned,
 } from '../../../state';
 import { trimWorkload } from '../../workload-sanitization';
+import { deleteWorkloadFromScanQueue } from './queue';
 
 export async function paginatedNamespacedReplicationControllerList(
   namespace: string,
@@ -80,6 +81,7 @@ export async function replicationControllerWatchHandler(
           .filter((container) => container.image !== undefined)
           .map((container) => container.image!),
       }),
+      deleteWorkloadFromScanQueue(workloadAlreadyScanned),
     ]);
   }
